catDog/requirement.txt
django
djangorestframework
numpy
pandas
scikit-learn